+++ src/display/style.rs
@@ -7,6 +7,7 @@ use line_numbers::SingleLineSpan;
 use owo_colors::{OwoColorize, Style};
 use unicode_width::{UnicodeWidthChar, UnicodeWidthStr};
 
+use crate::lines::split_on_newlines;
 use crate::parse::syntax::StringKind;
 use crate::{
     constants::Side,
@@ -91,7 +92,19 @@ fn split_string_by_width(s: &str, max_width: usize, tab_width: usize) -> Vec<(&s
     let mut parts: Vec<(&str, usize)> = vec![];
     let mut s = s;
 
-    while width_respecting_tabs(s, tab_width) > max_width {
+    // Optimisation: width_respecting_tabs() walks the whole string,
+    // which is slow when we have files with massive lines.
+    //
+    // A single character (grapheme) in UTF-8 can be 1, 2, 3 or 4
+    // bytes. A character's display width can be 0 (control
+    // characters), 1 (the typical case), 2 (e.g. fullwidth characters
+    // in Chinese, Japanese and Korean) or 4 (the default width for
+    // tabs in difftastic).
+    //
+    // Ignoring control characters, this means an n-byte UTF-8 string
+    // has a display width of at least n/4 characters. Check that case
+    // first, because it's a cheap conservative calculation.
+    while s.len() / 4 > max_width || width_respecting_tabs(s, tab_width) > max_width {
         let offset = byte_offset_for_width(s, max_width, tab_width);
 
         let part = substring_by_byte(s, 0, offset);
@@ -401,7 +414,7 @@ pub(crate) fn apply_colors(
     positions: &[MatchedPos],
 ) -> Vec<String> {
     let styles = color_positions(side, background, syntax_highlight, file_format, positions);
-    let lines = s.lines().collect::<Vec<_>>();
+    let lines = split_on_newlines(s).collect::<Vec<_>>();
     style_lines(&lines, &styles)
 }